code refact, new menu, new save load system, network connect by fqdn

+++ current/_replays.pas
@@ -8,7 +8,7 @@ vr,t,hp,tm:byte;
      mw:integer;
      wr:cardinal;
 begin
-   rpls_str_info:='';
+   rpls_str_info1:='';
 
    if(rpls_list_sel<0)or(rpls_list_sel>=rpls_list_size)then exit;
 
@@ -23,13 +23,13 @@ begin
       {$I+}
       if(ioresult<>0)then
       begin
-         rpls_str_info:=str_error_OpenFile;
+         rpls_str_info1:=str_error_OpenFile;
          exit;
       end;
       if(FileSize(f)<rpls_file_head_size)then
       begin
          close(f);
-         rpls_str_info:=str_error_WrongData;
+         rpls_str_info1:=str_error_WrongData;
          exit;
       end;
       vr:=0;
@@ -40,37 +40,41 @@ begin
          vr:=0;
          mw:=0;
          wr:=0;
-         BlockRead(f,wr,sizeof(map_seed    ));rpls_str_info:=str_map+': '+c2s(wr)+tc_nl2+' ';wr:=0;
-         BlockRead(f,mw,SizeOf(map_size    ));rpls_str_info+=str_map_size+i2s(mw)+tc_nl2+' ';mw:=0;
+         rpls_str_info1:=str_menu_map+tc_nl2+' ';
+         rpls_str_info2:=rpls_list[rpls_list_sel]+tc_nl2;
+
+         BlockRead(f,wr,sizeof(map_seed    ));rpls_str_info1+=str_map_seed+c2s(wr)+tc_nl2+' ';wr:=0;
+         BlockRead(f,mw,SizeOf(map_size    ));rpls_str_info1+=str_map_size+i2s(mw)+tc_nl2+' ';mw:=0;
          BlockRead(f,vr,sizeof(map_type    ));
-         if(vr>gms_m_types)then begin rpls_str_info:=str_error_WrongVersion;close(f);exit;end
-                           else       rpls_str_info+=str_map_type+str_map_typel[vr]+tc_default+tc_nl2+' '; vr:=0;
+         if(vr>gms_m_types)then begin rpls_str_info1:=str_error_WrongVersion;close(f);exit;     end
+                           else       rpls_str_info1+=str_map_type+str_map_typel[vr]+tc_default+tc_nl2+' '; vr:=0;
          BlockRead(f,vr,sizeof(map_symmetry));
-         if(vr>gms_m_symm )then begin rpls_str_info:=str_error_WrongVersion;close(f);exit;     end
-                           else       rpls_str_info+=str_map_sym+str_map_syml[vr]+tc_nl2+' '; vr:=0;
+         if(vr>gms_m_symm )then begin rpls_str_info1:=str_error_WrongVersion;close(f);exit;     end
+                           else       rpls_str_info1+=str_map_sym+str_map_syml[vr]+tc_nl2+' '; vr:=0;
 
          BlockRead(f,vr,sizeof(g_mode      ));
-         if(vr in allgamemodes)then begin rpls_str_info+=str_emnu_GameModel[vr]+tc_nl2;              end
-                               else begin rpls_str_info:=str_error_WrongVersion;close(f);exit;end;
+         if(vr in allgamemodes)then begin rpls_str_info1+=str_emnu_GameModel[vr]+tc_nl2;        end
+                               else begin rpls_str_info1:=str_error_WrongVersion;close(f);exit; end;
          BlockRead(f,vr,sizeof(g_start_base     ));vr:=0;
          BlockRead(f,vr,sizeof(g_fixed_positions));vr:=0;
          BlockRead(f,vr,sizeof(g_generators     ));vr:=0;
          BlockRead(f,hp,SizeOf(PlayerClient     ));
 
-         rpls_str_info+=tc_nl2;
+         rpls_str_info1+=tc_nl2+str_menu_players+tc_nl2;
 
          for vr:=1 to MaxPlayers do
          begin
             BlockRead(f,fn,sizeof(fn));   // name
 
-            rpls_str_info+=chr(vr)+pl_n_ch[vr=hp]+tc_default;
+            rpls_str_info1+=chr(vr)+pl_n_ch[vr=hp]+tc_default;
+            if(vr=hp)then rpls_str_info2+=pl_n_ch[vr=hp]+fn;
 
             t :=0;
             tm:=0;
             BlockRead(f,t ,1);  // state
             if(t=pt_none)then
             begin
-               rpls_str_info+=fn+tc_nl2;
+               rpls_str_info1+=fn;
                BlockRead(f,mw,5);
             end
             else
@@ -81,19 +85,20 @@ begin
                BlockRead(f,tm,1); // slot_state
 
                if(t=0)
-               then rpls_str_info+=str_observer[1]
+               then rpls_str_info1+=str_observer[1]
                else
                  if(tm<=r_cnt)
-                 then rpls_str_info+=str_racel[tm][2]
-                 else rpls_str_info+='?';
-               rpls_str_info+=','+t2c(t)+','+fn+tc_nl2;
+                 then rpls_str_info1+=str_racel[tm][2]
+                 else rpls_str_info1+='?';
+               rpls_str_info1+=','+t2c(t)+','+fn;
             end;
+            if(vr<MaxPlayers)then rpls_str_info1+=tc_nl2;
          end;
       end
-      else rpls_str_info:=str_error_WrongVersion;
+      else rpls_str_info1:=str_error_WrongVersion;
       close(f);
    end
-   else rpls_str_info:=str_error_FileExists;
+   else rpls_str_info1:=str_error_FileExists;
 end;
 
 procedure replay_CalcHeaderSize;
@@ -109,7 +114,7 @@ begin
                   +SizeOf(g_start_base     )
                   +SizeOf(g_fixed_positions)
                   +SizeOf(g_generators     )
-                  +sizeof(rpls_player      );
+                  +sizeof(rpls_POVPlayer   );
    with g_players[0] do
    rpls_file_head_size
                  +=cardinal((sizeof(name       )
@@ -206,6 +211,35 @@ begin
    setlength(rpls_ReadPosl,rpls_ReadPosN);
 end;
 
+
+function str_DateTime:shortstring;
+var YY,MM,DD,H,M,S,MS:word;
+begin
+   DeCodeDate(Date,YY,MM,DD);
+   DeCodeTime(Time,H,M,S,MS);
+   str_DateTime:=w2s(YY)+'_'+w2s(MM)+'_'+w2s(DD)+' '+w2s(H)+'-'+w2s(M)+'-'+w2s(S)+'-'+w2s(MS);
+end;
+
+
+procedure replay_MakeNewReplayPath;
+const
+replay_template_n = '%N%';
+var    p: byte;
+       i: integer;
+tmp_strt: shortstring;
+begin
+   // rpls_str_path:=str_f_rpls+rpls_str_name+str_e_rpls;
+   tmp_strt:='';
+
+{
+i:=0;
+repeat
+   i+=1;
+   s:=str_screenshot+i2s(i)+'.bmp';
+until not FileExists(s);
+}
+end;
+
 procedure replay_Code;
 const vxyc = 5;
 var  i,gs,
@@ -228,21 +262,22 @@ begin
    begin
       replay_Abort;
       rpls_state:=rpls_state_none;
+      GameLogChat(PlayerClient,log_to_all,str_msg_ReplayFail+str_error_OpenFile,true);
    end
    else
    begin
-      rpls_fstatus:=rpls_file_write;
-      rpls_state  :=rpls_state_write;
-      rpls_u      :=MaxPlayerUnits+1;
-      rpls_player :=PlayerClient;
-      rpls_log_n  :=g_players[rpls_player].log_n;
-      rpls_plcam  :=false;
-      rpls_ticks  :=0;
+      rpls_fstatus  :=rpls_file_write;
+      rpls_state    :=rpls_state_write;
+      rpls_u        :=MaxPlayerUnits+1;
+      rpls_POVPlayer:=PlayerClient;
+      rpls_log_n    :=g_players[rpls_POVPlayer].log_n;
+      rpls_plcam    :=false;
+      rpls_ticks    :=0;
 
       {$I-}
       BlockWrite(rpls_file,g_version        ,SizeOf(g_version        ));
       BlockWrite(rpls_file,map_seed         ,SizeOf(map_seed         ));
-      BlockWrite(rpls_file,map_size           ,SizeOf(map_size           ));
+      BlockWrite(rpls_file,map_size         ,SizeOf(map_size         ));
       BlockWrite(rpls_file,map_type         ,SizeOf(map_type         ));
       BlockWrite(rpls_file,map_symmetry     ,SizeOf(map_symmetry     ));
 
@@ -250,7 +285,7 @@ begin
       BlockWrite(rpls_file,g_start_base     ,SizeOf(g_start_base     ));
       BlockWrite(rpls_file,g_fixed_positions,SizeOf(g_fixed_positions));
       BlockWrite(rpls_file,g_generators     ,SizeOf(g_generators     ));
-      BlockWrite(rpls_file,rpls_player      ,sizeof(rpls_player      ));
+      BlockWrite(rpls_file,rpls_POVPlayer   ,sizeof(rpls_POVPlayer   ));
       {$I+}
 
       for p:=1 to MaxPlayers do
@@ -270,6 +305,7 @@ begin
       begin
          replay_Abort;
          rpls_state:=rpls_state_none;
+         GameLogChat(PlayerClient,log_to_all,str_msg_ReplayFail+str_error_FileWrite,true);
       end;
    end;
 end;
@@ -282,7 +318,7 @@ begin
 
    gs:=G_Status and %00111111;
    i :=gs;
-   if(g_players[rpls_player].log_n<>rpls_log_n)then i:=i or %10000000;
+   if(g_players[rpls_POVPlayer].log_n<>rpls_log_n)then i:=i or %10000000;
    if(rpls_vidx<>_vx)or(rpls_vidy<>_vy)then
     if(gs=gs_running)then i:=i or %01000000;
 
@@ -291,7 +327,7 @@ begin
       {$I-}
       BlockWrite(rpls_file,i,sizeof(i));
       {$I+}
-      if((i and %10000000)>0)then wudata_log(rpls_player,@rpls_log_n,true);
+      if((i and %10000000)>0)then wudata_log(rpls_POVPlayer,@rpls_log_n,true);
       if((i and %01000000)>0)then
       begin
          rpls_vidx:=_vx;
@@ -302,13 +338,14 @@ begin
          {$I+}
       end;
 
-      if(gs=gs_running)then wclinet_gframe(rpls_player,true);
+      if(gs=gs_running)then wclinet_gframe(rpls_POVPlayer,true);
    end;
 
    if(ioresult<>0)then
    begin
       replay_Abort;
       rpls_state:=rpls_state_none;
+      GameLogChat(PlayerClient,log_to_all,str_msg_ReplayFail+str_error_FileWrite,true);
    end;
 end;
 
@@ -320,9 +357,10 @@ begin
 
    if(rpls_list_sel<0)or(rpls_list_sel>=rpls_list_size)then
    begin
-      rpls_state   :=rpls_state_none;
-      g_started    :=false;
-      rpls_str_info:='';
+      rpls_state    :=rpls_state_none;
+      G_Started     :=false;
+      rpls_str_info1:='';
+      rpls_str_info2:='';
       exit;
    end;
 
@@ -331,8 +369,9 @@ begin
    if(not FileExists(rpls_str_path))then
    begin
       rpls_state   :=rpls_state_none;
-      g_started    :=false;
-      rpls_str_info:=str_error_FileExists;
+      G_Started    :=false;
+      rpls_str_info1:=str_error_FileExists;
+      //GameLogChat(PlayerClient,log_to_all,str_error_FileExists+' '+rpls_list[rpls_list_sel],true);
       exit;
    end;
 
@@ -344,8 +383,9 @@ begin
    if(ioresult<>0)then
    begin
       replay_Abort;
-      g_started    :=false;
-      rpls_str_info:=str_error_OpenFile;
+      G_Started    :=false;
+      rpls_str_info1:=str_error_OpenFile;
+      //GameLogChat(PlayerClient,log_to_all,str_error_OpenFile+' '+rpls_list[rpls_list_sel],true);
    end
    else
    begin
@@ -354,8 +394,8 @@ begin
       if(rpls_file_size<rpls_file_head_size)then
       begin
          replay_Abort;
-         g_started    :=false;
-         rpls_str_info:=str_error_WrongData;
+         G_Started    :=false;
+         rpls_str_info1:=str_error_WrongData;
          exit;
       end;
 
@@ -367,8 +407,8 @@ begin
       if(i<>g_version)then
       begin
          replay_Abort;
-         g_started    :=false;
-         rpls_str_info:=str_error_WrongVersion;
+         G_Started    :=false;
+         rpls_str_info1:=str_error_WrongVersion;
       end
       else
       begin
@@ -383,20 +423,20 @@ begin
          BlockRead(rpls_file,g_start_base     ,SizeOf(g_start_base     ));
          BlockRead(rpls_file,g_fixed_positions,SizeOf(g_fixed_positions));
          BlockRead(rpls_file,g_generators     ,SizeOf(g_generators     ));
-         BlockRead(rpls_file,rpls_player      ,sizeof(rpls_player      ));
+         BlockRead(rpls_file,rpls_POVPlayer   ,sizeof(rpls_POVPlayer   ));
          {$I+}
 
          if(map_size<MinMapSize)or(map_size>MaxMapSize)
-         or(map_type    >gms_m_types)
-         or(map_symmetry>gms_m_symm)
-         or(g_mode>gms_count)
-         or(g_start_base>gms_g_startb )
-         or(g_generators>gms_g_maxgens)
-         or(rpls_player>MaxPlayers)then
+         or(map_type      >gms_m_types  )
+         or(map_symmetry  >gms_m_symm   )
+         or(g_mode        >gms_count    )
+         or(g_start_base  >gms_g_startb )
+         or(g_generators  >gms_g_maxgens)
+         or(rpls_POVPlayer>MaxPlayers   )then
          begin
             replay_Abort;
-            g_started:=false;
-            rpls_str_info:=str_error_WrongVersion;
+            G_Started:=false;
+            rpls_str_info1:=str_error_WrongVersion;
             GameDefaultAll;
             exit;
          end;
@@ -418,8 +458,9 @@ begin
          if(ioresult<>0)then
          begin
             replay_Abort;
-            rpls_str_info:=str_error_WrongVersion;
+            rpls_str_info1:=str_error_WrongVersion;
             GameDefaultAll;
+            //GameLogChat(PlayerClient,log_to_all,str_error_WrongVersion+' '+rpls_list[rpls_list_sel],true);
             exit;
          end;
 
@@ -433,7 +474,7 @@ begin
          rpls_state  :=rpls_state_read;
          rpls_pnu    :=0;
          rpls_ticks  :=0;
-         PlayerClient:=rpls_player;
+         PlayerClient:=rpls_POVPlayer;
          UIPlayer    :=PlayerClient;
 
          rpls_plcam  :=false;
@@ -464,9 +505,9 @@ begin
 
    if(eof(rpls_file))then
    begin
-      G_Status   :=gs_replayend;
-      sys_uncappedFPS:=false;
-      rpls_ForwardStep  :=0;
+      G_Status        :=gs_replayend;
+      sys_uncappedFPS :=false;
+      rpls_ForwardStep:=0;
       exit;
    end;
 
@@ -482,7 +523,7 @@ begin
       {$I+}
       G_Status:=i and %00111111;
 
-      if((i and %10000000)>0)then rudata_log(rpls_player,true);
+      if((i and %10000000)>0)then rudata_log(rpls_POVPlayer,true);
       if((i and %01000000)>0)then
       begin
          {$I-}
@@ -491,7 +532,7 @@ begin
          {$I+}
       end;
 
-      if(G_Status=gs_running)then rclinet_gframe(rpls_player,true,rpls_ForwardStep>1);
+      if(G_Status=gs_running)then rclinet_gframe(rpls_POVPlayer,true,rpls_ForwardStep>1);
 
       if(rpls_ForwardStep>1)then effects_sprites(false,false);
       rpls_ForwardStep-=1;
@@ -508,20 +549,32 @@ begin
 end;
 
 begin
-   rpls_ticks+=1;
-   if(G_Started=false)or(rpls_state=rpls_state_none)//or(menu_s2=ms2_camp)
+   if(rpls_StartRecordPause>0)
+   then rpls_StartRecordPause-=1
+   else
+     if(rpls_Recording)and(G_Started)and(rpls_state=rpls_state_none)then
+     begin
+        rpls_state:=rpls_state_write;
+        rpls_StartRecordPause:=fr_fps2;
+        GameLogChat(PlayerClient,log_to_all,str_msg_ReplayStart+rpls_str_name,true);
+     end;
+
+   if(not G_Started)or(rpls_state=rpls_state_none)//or(menu_s2=ms2_camp)
    then replay_Abort
    else
      if(G_Started)then
-       case rpls_state of
+     begin
+        rpls_ticks+=1;
+        case rpls_state of
 rpls_state_write : if(rpls_fstatus<>rpls_file_write)
                    then replay_WriteHead
                    else replay_WriteGameFrame;
 rpls_state_read  : if(rpls_fstatus<>rpls_file_read)
                    then replay_Readhead
                    else replay_ReadGameFrame;
-       else replay_Abort;
-       end;
+        else replay_Abort;
+        end;
+     end;
 end;
 
 procedure replay_Select;
@@ -529,7 +582,7 @@ begin
    if(0<=rpls_list_sel)and(rpls_list_sel<rpls_list_size)
    then replay_MenuSelectedInfo
    else
-     if(g_started=false)then rpls_str_info:='';
+     if(G_Started=false)then rpls_str_info1:='';
 end;
 
 procedure replay_MakeFolderList;
@@ -537,7 +590,7 @@ var Info : TSearchRec;
        s : shortstring;
 begin
    rpls_list_scroll:=0;
-   rpls_list_size:=0;
+   rpls_list_size  :=0;
    setlength(rpls_list,0);
    if(FindFirst(str_f_rpls+'*'+str_e_rpls,faReadonly,info)=0)then
     repeat
@@ -555,14 +608,31 @@ begin
    replay_Select;
 end;
 
-procedure replay_Delete;
+function replay_Play(Check:boolean):boolean;
+begin
+   replay_Play:=false;
+   if(rpls_list_sel<0)or(rpls_list_sel>=rpls_list_size)then exit;
+
+   replay_Play:=true;
+   if(Check)then exit;
+
+   rpls_state:=rpls_state_read;
+   G_Started :=true;
+   menu_page1:=mp_scirmish;
+end;
+
+function replay_Delete(Check:boolean):boolean;
 var fn:shortstring;
 begin
+   replay_Delete:=false;
    if(rpls_list_sel<0)or(rpls_list_sel>=rpls_list_size)then exit;
 
    fn:=str_f_rpls+rpls_list[rpls_list_sel]+str_e_rpls;
    if(FileExists(fn))then
    begin
+      replay_Delete:=true;
+      if(Check)then exit;
+
       DeleteFile(fn);
       replay_MakeFolderList;
    end;
